ScrapingFood/layer_1/requirements.txt
pandas==2.0.2
requests==2.28.2
beautifulsoup4==4.12.0
bs4==0.0.1
deepl==1.14.0
python-dotenv==1.0.0
lxml==4.9.2